+++ R/docx_exporter_functions.R
@@ -3,6 +3,23 @@ dps_markup_df_docx <- tibble::tibble(
   replace_by = c("flextable::as_sup", "flextable::as_sub")
 )
 
+align_rows_with_rtf <- function(doc) {
+  # when we have vertical pagination in the case of tables and listings,
+  # from page 2 onwards, the rows visually look a little bit shifted downwards
+  # in RTF compared to DOCX.
+  # It's a tiny difference caused by a little horizontal spacing located
+  # between the repeated table header and the first body row in each page,
+  # starting from page 2.
+  # This visual difference has to do with file "settings.xml" in the docx,
+  # and more precisely with this node:
+  # <w:compat>
+  #   <w:compatSetting w:name="compatibilityMode" w:uri="http://schemas.microsoft.com/office/word" w:val="15"/>
+  # </w:compat>
+  # we need to change the value from 15 to 11
+  doc$settings$compatibility_mode <- "11"
+  return(doc)
+}
+
 validate_tabletree <- function(tt, validate, tlgtype) {
   if (validate && tlgtype == "Table" && methods::is(tt, "VTableTree")) {
     if (!rtables::validate_table_struct(tt)) {
@@ -82,7 +99,7 @@ export_as_csv <- function(tlgtype, export_csv, pags, fontspec,
 }
 
 
-insert_fake_watermark_XML <- function(doc, watermark, orientation) {
+insert_fake_watermark_XML <- function(doc, watermark, orientation, tlgtype) {
   # This function inserts a fake watermark WordArt in the Title of
   # the flextable (that is, in the document body) by inserting an XML node.
   # This function should only be called when exporting Figures.
@@ -93,13 +110,13 @@ insert_fake_watermark_XML <- function(doc, watermark, orientation) {
   checkmate::assert_choice(orientation, choices = c("portrait", "landscape"))
 
   if (orientation == "portrait") {
-    margin_left <- -118.05
-    margin_top <- 265.35
+    margin_left <- -114
+    margin_top <- 266
   } else {
     # to shift to the right, increase margin_left
     # to shift upwards, decrease margin_top
-    margin_left <- -30.05
-    margin_top <- 175.35
+    margin_left <- -24
+    margin_top <- ifelse(tlgtype == "Listing", 165, 176)
   }
 
 
@@ -372,7 +389,7 @@ add_hanging_indent_in_title_XML <- function(doc) {
 
 add_little_gap_bottom_borders_spanning_headers <- function(
   flx,
-  border = flextable::fp_border_default(width = 0.75, color = "black")
+  border = flextable::fp_border_default(width = 0.875, color = "black")
 ) {
   nrow_header <- flextable::nrow_part(x = flx, part = "header")
   if (nrow_header > 1) {
@@ -444,7 +461,7 @@ insert_footer_text <- function(flx, tblid) {
 
 insert_title_as_header <- function(flx,
                                    title,
-                                   border = flextable::fp_border_default(width = 0.75, color = "black")) {
+                                   border = flextable::fp_border_default(width = 0.875, color = "black")) {
   # this function inserts the Title as a header but does not attempt
   # to simulate the hanging indent. Instead, it adds the string as is, and the hanging indent
   # will be handled further downstream when exporting to docx by manipulating the XML.
@@ -642,7 +659,7 @@ theme_docx_default_j <- function(
   cell_margins = c(0, 0, 0, 0),
   bold = c("header", "content_rows", "label_rows", "top_left"),
   bold_manual = NULL,
-  border = flextable::fp_border_default(width = 0.75, color = "black")
+  border = flextable::fp_border_default(width = 0.875, color = "black")
 ) {
   function(flx, ...) {
     if (!inherits(flx, "flextable")) {
@@ -785,9 +802,9 @@ theme_docx_default_j <- function(
 #' @export
 tt_to_flextable_j <- function(
     tt,
-    tblid,
+    tblid = NULL,
     theme = theme_docx_default_j(font = "Times New Roman", font_size = 9L, bold = NULL),
-    border = flextable::fp_border_default(width = 0.75, color = "black"),
+    border = flextable::fp_border_default(width = 0.875, color = "black"),
     titles_as_header = TRUE,
     bold_titles = TRUE,
     integrate_footers = TRUE,
@@ -820,7 +837,7 @@ tt_to_flextable_j <- function(
   }
 
   checkmate::assert_true(inherits(tt, "VTableTree") || inherits(tt, "listing_df"))
-  checkmate::assert_character(tblid)
+  checkmate::assert_character(tblid, null.ok = TRUE)
   checkmate::assert_function(theme)
   checkmate::assert_class(border, "fp_border")
   checkmate::assert_flag(titles_as_header)
@@ -849,6 +866,11 @@ tt_to_flextable_j <- function(
   stopifnot("`alignments` must be a list" = is.list(alignments))
   for (alignment in alignments) {
     stopifnot("Each item of `alignments` must be a list" = is.list(alignment))
+  }
+
+  if (is.null(tblid)) {
+    temp_file <- tempfile()
+    tblid <- basename(temp_file)
   }
 
   validate_tabletree(tt, validate, tlgtype)
@@ -1000,7 +1022,9 @@ tt_to_flextable_j <- function(
         # only if for each vertical pagination only the first row has indentation == 0
         only_first_row_indent_zero <-
           all(lapply(full_pag_i, function(x) {
-            x$row_info |> dplyr::filter(label != " ", indent == 0) |> nrow()
+            tmp <- x$row_info
+            tmp <- tmp[tmp$label != " " & tmp$indent == 0, ]
+            return(nrow(tmp))
           }) == 1)
 
         sub_ft <- tt_to_flextable_j(
@@ -1477,8 +1501,8 @@ tt_to_flextable_j <- function(
 #' @inheritSection export_TLG_as_docx Note
 export_as_docx_j <- function(
   tt,
-  tblid,
-  output_dir,
+  tblid = NULL,
+  output_dir = NULL,
   theme = theme_docx_default_j(
     font = "Times New Roman",
     font_size = 9L,
@@ -1507,7 +1531,7 @@ export_as_docx_j <- function(
   pagenum = ifelse(tlgtype == "Listing", TRUE, FALSE),
   round_type = formatters::obj_round_type(tt),
   alignments = list(),
-  border = flextable::fp_border_default(width = 0.75, color = "black"),
+  border = flextable::fp_border_default(width = 0.875, color = "black"),
   border_mat = make_header_bordmat(obj = tt),
   watermark = NULL,
   export_csv = FALSE,
@@ -1519,9 +1543,11 @@ export_as_docx_j <- function(
 
   checkmate::assert_true(inherits(tt, "VTableTree") || inherits(tt, "listing_df") ||
                            is.list(tt) || inherits(tt, "flextable"))
-  checkmate::assert_character(tblid)
-  checkmate::assert_character(output_dir, len = 1)
-  checkmate::assert_directory_exists(output_dir)
+  checkmate::assert_character(tblid, null.ok = TRUE)
+  checkmate::assert_character(output_dir, len = 1, null.ok = TRUE)
+  if (!is.null(output_dir)) {
+    checkmate::assert_directory_exists(output_dir)
+  }
   checkmate::assert_function(theme)
   checkmate::assert_flag(add_page_break)
   checkmate::assert_flag(titles_as_header)
@@ -1559,6 +1585,12 @@ export_as_docx_j <- function(
   checkmate::assert_character(watermark, len = 1, null.ok = TRUE)
   checkmate::assert_flag(export_csv)
   checkmate::assert_character(output_csv_directory, null.ok = TRUE, len = 1)
+
+  if (is.null(tblid) || is.null(output_dir)) {
+    temp_file <- tempfile()
+    tblid <- basename(temp_file)
+    output_dir <- dirname(temp_file)
+  }
 
   do_tt_error <- FALSE
   if (tlgtype != "Listing") {
@@ -1798,7 +1830,7 @@ export_as_docx_j <- function(
     if (isFALSE(integrate_footers) && inherits(tt, "VTableTree")) {
       matform <- rtables::matrix_form(tt, indent_rownames = TRUE, round_type = round_type)
       if (length(matform$ref_footnotes) > 0) {
-        char_v <- matform$ref_footnotes
+        chr_v <- matform$ref_footnotes
         text_format <- flx_fpt$fpt_footer
         for (ii in seq_along(chr_v)) {
           cur_fp <- officer::fpar(officer::ftext(chr_v[ii], prop = text_format))
@@ -1806,7 +1838,7 @@ export_as_docx_j <- function(
         }
       }
       if (length(formatters::all_footers(tt)) > 0) {
-        char_v <- formatters::all_footers(tt)
+        chr_v <- formatters::all_footers(tt)
         text_format <- flx_fpt$fpt_footer
         for (ii in seq_along(chr_v)) {
           cur_fp <- officer::fpar(officer::ftext(chr_v[ii], prop = text_format))
@@ -1829,8 +1861,10 @@ export_as_docx_j <- function(
     add_vertical_pagination_XML(doc)
     remove_security_popup_page_numbers_XML(doc, tlgtype, pagenum)
     if (!is.null(watermark)) {
-      insert_fake_watermark_XML(doc, watermark, orientation)
+      insert_fake_watermark_XML(doc, watermark, orientation, tlgtype)
     }
+
+    doc <- align_rows_with_rtf(doc)
 
     print(doc, target = paste0(output_dir, "/", tolower(tblid), ".docx"))
     invisible(TRUE)
@@ -1859,15 +1893,15 @@ export_as_docx_j <- function(
 #' @inheritSection export_TLG_as_docx Note
 export_graph_as_docx <- function(g = NULL,
                                  plotnames = NULL,
-                                 tblid,
-                                 output_dir,
+                                 tblid = NULL,
+                                 output_dir = NULL,
                                  title = NULL,
                                  footers = NULL,
                                  orientation = "portrait",
                                  plotwidth = 8,
                                  plotheight = 5.51,
                                  units = c("in", "cm", "mm", "px")[1],
-                                 border = flextable::fp_border_default(width = 0.75, color = "black"),
+                                 border = flextable::fp_border_default(width = 0.875, color = "black"),
                                  watermark = NULL) {
 
   # Check arguments ----
@@ -1922,9 +1956,11 @@ export_graph_as_docx <- function(g = NULL,
     }
   }
 
-  checkmate::assert_character(tblid)
-  checkmate::assert_character(output_dir, len = 1)
-  checkmate::assert_directory_exists(output_dir)
+  checkmate::assert_character(tblid, null.ok = TRUE)
+  checkmate::assert_character(output_dir, len = 1, null.ok = TRUE)
+  if (!is.null(output_dir)) {
+    checkmate::assert_directory_exists(output_dir)
+  }
   checkmate::assert_character(title, null.ok = TRUE)
   checkmate::assert_character(footers, null.ok = TRUE)
   checkmate::assert_choice(orientation, choices = c("portrait", "landscape"))
@@ -1933,6 +1969,12 @@ export_graph_as_docx <- function(g = NULL,
   checkmate::assert_choice(units, choices = c("in", "cm", "mm", "px"))
   checkmate::assert_class(border, "fp_border")
   checkmate::assert_character(watermark, len = 1, null.ok = TRUE)
+
+  if (is.null(tblid) || is.null(output_dir)) {
+    temp_file <- tempfile()
+    tblid <- basename(temp_file)
+    output_dir <- dirname(temp_file)
+  }
 
 
   # Creation of flextable ----
@@ -2030,7 +2072,7 @@ export_graph_as_docx <- function(g = NULL,
   doc <- flextable::body_add_flextable(doc, flx, align = "center")
   add_hanging_indent_in_title_XML(doc)
   if (!is.null(watermark)) {
-    insert_fake_watermark_XML(doc, watermark, orientation)
+    insert_fake_watermark_XML(doc, watermark, orientation, "Figure")
   }
   print(doc, target = paste0(output_dir, "/", tolower(tblid), ".docx"))
 }
@@ -2042,8 +2084,14 @@ export_graph_as_docx <- function(g = NULL,
 #'
 #' @param obj (`TableTree`, `listing_df` or `ggplot2`)\cr the object to export.
 #' @param tblid (`character`)\cr output ID to be displayed in the title and
-#' last line of footer.
-#' @param output_dir (`character`)\cr a directory path to save the docx.
+#' last line of footer. When exporting, it will also be used as the output filename.\cr
+#' If NULL, a temp file will be created, its dirname will replace argument `output_dir`,
+#' and its basename will replace argument `tblid`.\cr
+#' (optional) Default = NULL.
+#' @param output_dir (`character`)\cr a directory path to save the docx.\cr
+#' If NULL, a temp file will be created, its dirname will replace argument `output_dir`,
+#' and its basename will replace argument `tblid`.\cr
+#' (optional) Default = NULL.
 #' @param theme (function factory)\cr the theme to apply to the flextable.\cr
 #' (optional) Default = [theme_docx_default_j()].\cr
 #' See [theme_docx_default_j()] or [rtables.officer::theme_docx_default()]
@@ -2062,7 +2110,7 @@ export_graph_as_docx <- function(g = NULL,
 #' @param template_file (`character`)\cr Template file that `officer` will use as a starting
 #' point for the final document. Document attaches the table and uses the defaults
 #' defined in the template file. Paragraph styles are inherited from this file.\cr
-#' If NULL, this function will use an internal template.
+#' If NULL, this function will use an internal template.\cr
 #' (optional) Default = NULL.
 #' @param orientation (`character`)\cr one of: "portrait", "landscape".\cr
 #' (optional) Default = "portrait".
@@ -2091,7 +2139,7 @@ export_graph_as_docx <- function(g = NULL,
 #' Each named list contains `row`, `col`, and `value`.\cr
 #' (optional) Default = `list()`.
 #' @param border (`fp_border`)\cr border to use.\cr
-#' Default = \code{flextable::fp_border_default(width = 0.75, color = "black")}.
+#' Default = \code{flextable::fp_border_default(width = 0.875, color = "black")}.
 #' @param border_mat (`matrix`)\cr a `m x k` matrix where m is the number of columns of
 #' the input Table/Listing and k is the number of lines the header takes up.\cr
 #' See [tidytlg::add_bottom_borders] for what the matrix should contain.
@@ -2165,7 +2213,14 @@ export_graph_as_docx <- function(g = NULL,
 #'   titles_as_header = TRUE, integrate_footers = TRUE,
 #'   section_properties = officer::prop_section(
 #'     page_size = officer::page_size(width = 11, height = 8.5, orient = "portrait"),
-#'     page_margins = officer::page_mar(bottom = 1, top = 1, right = 1, left = 1, gutter = 0, footer = 1, header = 1)
+#'     page_margins = officer::page_mar(
+#'                              bottom = 1,
+#'                              top = 1,
+#'                              right = 1,
+#'                              left = 1,
+#'                              gutter = 0,
+#'                              footer = 1,
+#'                              header = 1)
 #'   ),
 #'   doc_metadata = NULL,
 #'   template_file = NULL,
@@ -2183,7 +2238,7 @@ export_graph_as_docx <- function(g = NULL,
 #'   pagenum = FALSE,
 #'   round_type = "iec",
 #'   alignments = list(),
-#'   border = flextable::fp_border_default(width = 0.75, color = "black"),
+#'   border = flextable::fp_border_default(width = 0.875, color = "black"),
 #'   border_mat = NULL,
 #'   watermark = NULL,
 #'   plotnames = NULL,
@@ -2206,8 +2261,8 @@ export_graph_as_docx <- function(g = NULL,
 #'
 export_TLG_as_docx <- function(
   obj = NULL,
-  tblid,
-  output_dir,
+  tblid = NULL,
+  output_dir = NULL,
   theme = theme_docx_default_j(
     font = "Times New Roman",
     font_size = 9L,
@@ -2238,7 +2293,7 @@ export_TLG_as_docx <- function(
                       formatters::obj_round_type(obj),
                       "iec"),
   alignments = list(),
-  border = flextable::fp_border_default(width = 0.75, color = "black"),
+  border = flextable::fp_border_default(width = 0.875, color = "black"),
   border_mat = NULL,
   export_csv = FALSE,
   output_csv_directory = NULL,
@@ -2265,9 +2320,11 @@ export_TLG_as_docx <- function(
     checkmate::assert_true(inherits(obj, "VTableTree") || inherits(obj, "listing_df") ||
                              ggplot2::is.ggplot(obj) || is.list(obj))
   }
-  checkmate::assert_character(tblid)
-  checkmate::assert_character(output_dir, len = 1)
-  checkmate::assert_directory_exists(output_dir)
+  checkmate::assert_character(tblid, null.ok = TRUE)
+  checkmate::assert_character(output_dir, len = 1, null.ok = TRUE)
+  if (!is.null(output_dir)) {
+    checkmate::assert_directory_exists(output_dir)
+  }
   checkmate::assert_function(theme)
   checkmate::assert_flag(add_page_break)
   checkmate::assert_flag(titles_as_header)