+++ tests/testthat/test-docx_exporter_functions.R
@@ -53,20 +53,20 @@ rtables::label_at_path(tbl1c, c("COUNTRY", "count_unique_denom_fraction.JPN")) <
 
 snapshot_test_docx <- function(doc, detailed = TRUE) {
   if (Sys.info()[["sysname"]] == "Windows") {
-    testthat::expect_snapshot(officer::docx_summary(x = doc, detailed = detailed))
+    testthat::expect_snapshot(cran = TRUE, officer::docx_summary(x = doc, detailed = detailed))
   }
 }
 
 snapshot_test_flextable <- function(res) {
   testthat::expect_true(inherits(res, "flextable"))
   if (Sys.info()[["sysname"]] == "Windows") {
-    testthat::expect_snapshot(res$header)
-    testthat::expect_snapshot(res$body)
-    testthat::expect_snapshot(res$footer)
-    testthat::expect_snapshot(res$col_keys)
-    testthat::expect_snapshot(res$caption)
-    testthat::expect_snapshot(res$blanks)
-    testthat::expect_snapshot(res$properties)
+    testthat::expect_snapshot(cran = TRUE, res$header)
+    testthat::expect_snapshot(cran = TRUE, res$body)
+    testthat::expect_snapshot(cran = TRUE, res$footer)
+    testthat::expect_snapshot(cran = TRUE, res$col_keys)
+    testthat::expect_snapshot(cran = TRUE, res$caption)
+    testthat::expect_snapshot(cran = TRUE, res$blanks)
+    testthat::expect_snapshot(cran = TRUE, res$properties)
 
     doc <- officer::read_docx()
     doc <- flextable::body_add_flextable(doc, res, align = "center")
@@ -133,7 +133,7 @@ testthat::test_that("tt_to_flextable_j() works fine with border_mat", {
   flx2 <- tt_to_flextable_j(tt = tbl2, tblid = "output ID", border_mat = border_mat)
 
   expected_res <- flx1$header$styles$cells$border.width.bottom$data
-  expected_res[2, 4] <- 0.75
+  expected_res[2, 4] <- 0.875
   testthat::expect_equal(flx2$header$styles$cells$border.width.bottom$data, expected_res)
 
 })