+++ tests/testthat/test-mmrm.R
@@ -25,7 +25,7 @@ test_that("get_mmrm_lsmeans can calculate the LS mean results including one- and
     weights = weights
   )
 
-  expect_snapshot(result)
+  expect_snapshot(cran = TRUE, result)
 
   # Additional p-value verification
   contrasts <- result$contrasts
@@ -37,6 +37,39 @@ test_that("get_mmrm_lsmeans can calculate the LS mean results including one- and
   expect_equal(contrasts$p_value, two_sided_pvals)
   expect_equal(contrasts$p_value_less, pvals[, "lower"])
   expect_equal(contrasts$p_value_greater, pvals[, "upper"])
+})
+
+test_that("get_mmrm_lsmeans can adjust contrasts for multiplicity within visits", {
+  vars <- list(
+    response = "FEV1",
+    id = "USUBJID",
+    arm = "RACE",
+    visit = "AVISIT"
+  )
+  fit <- mmrm::mmrm(
+    formula = FEV1 ~ RACE * AVISIT + us(AVISIT | USUBJID),
+    data = mmrm::fev_data
+  )
+  conf_level <- 0.95
+  weights <- "counterfactual"
+
+  result_dunnett <- get_mmrm_lsmeans(
+    fit = fit,
+    vars = vars,
+    conf_level = conf_level,
+    weights = weights,
+    mult_adj = "dunnett"
+  )
+  expect_snapshot_value(result_dunnett, style = "deparse")
+
+  result_step_down_dunnett <- get_mmrm_lsmeans(
+    fit = fit,
+    vars = vars,
+    conf_level = conf_level,
+    weights = weights,
+    mult_adj = "step-down-dunnett"
+  )
+  expect_snapshot_value(result_step_down_dunnett, style = "deparse")
 })
 
 test_that("fit_mmrm_j works as expected", {
@@ -56,5 +89,5 @@ test_that("fit_mmrm_j works as expected", {
     )
   )
 
-  expect_snapshot(fit)
+  expect_snapshot(cran = TRUE, fit)
 })